requirements.txt
boto3==1.34.3
botocore==1.34.3
jmespath==1.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.9.0
six==1.16.0
slack-sdk==3.26.1
urllib3==2.1.0
